moved offsetLineCommands to UIValidator and made it private
